Expand gpu features flags to 64-bit.

+++ GPU/GLES/GPU_GLES.cpp
@@ -491,7 +491,7 @@ GPU_GLES::~GPU_GLES() {
 
 // Take the raw GL extension and versioning data and turn into feature flags.
 void GPU_GLES::CheckGPUFeatures() {
-	u32 features = 0;
+	u64 features = 0;
 	if (gl_extensions.ARB_blend_func_extended || gl_extensions.EXT_blend_func_extended) {
 		if (gl_extensions.gpuVendor == GPU_VENDOR_INTEL || !gl_extensions.VersionGEThan(3, 0, 0)) {
 			// Don't use this extension to off on sub 3.0 OpenGL versions as it does not seem reliable
